% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_oxide.R
\name{pdf_is_encrypted}
\alias{pdf_is_encrypted}
\title{Whether the document is encrypted.}
\usage{
pdf_is_encrypted(doc)
}
\arguments{
\item{doc}{A `pdfoxide_document`.}
}
\description{
Whether the document is encrypted.
}
