% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_oxide.R
\name{pdf_open_bytes}
\alias{pdf_open_bytes}
\title{Open a PDF document from a raw vector.}
\usage{
pdf_open_bytes(bytes)
}
\arguments{
\item{bytes}{A `raw` vector.}
}
\value{
A `pdfoxide_document` handle.
}
\description{
Open a PDF document from a raw vector.
}
