% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_oxide.R
\name{pdf_save}
\alias{pdf_save}
\title{Save a built PDF to a path.}
\usage{
pdf_save(pdf, path)
}
\arguments{
\item{pdf}{A `pdfoxide_pdf`. @param path Output path.}
}
\description{
Save a built PDF to a path.
}
