% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_oxide.R
\name{pdf_to_markdown_all}
\alias{pdf_to_markdown_all}
\title{Markdown for the whole document.}
\usage{
pdf_to_markdown_all(doc)
}
\arguments{
\item{doc}{A `pdfoxide_document`.}
}
\description{
Markdown for the whole document.
}
